chore: Update farm rpc calls
